[修正] ユーザ認証
