+++ src/app/src/UI/userAuth/SignUp.tsx
@@ -1,50 +1,28 @@
 import * as React from "react";
-import Avatar from "@mui/material/Avatar";
-import Button from "@mui/material/Button";
-import CssBaseline from "@mui/material/CssBaseline";
-import TextField from "@mui/material/TextField";
-import Grid from "@mui/material/Grid";
-import Box from "@mui/material/Box";
-import LockOutlinedIcon from "@mui/icons-material/LockOutlined";
-import Typography from "@mui/material/Typography";
-import Container from "@mui/material/Container";
-import { createTheme, ThemeProvider } from "@mui/material/styles";
-import { auth } from "../firebase";
 import { Link, useNavigate } from "react-router-dom";
 
-// function Copyright(props: any) {
-//   return (
-//     <Typography variant="body2" color="text.secondary" align="center" {...props}>
-//       {'Copyright © '}
-//       <Link to="/">
-//         Your Website
-//       </Link>{' '}
-//       {new Date().getFullYear()}
-//       {'.'}
-//     </Typography>
-//   );
-// }
+import Avatar from "@mui/material/Avatar";
+import Box from "@mui/material/Box";
+import Button from "@mui/material/Button";
+import Container from "@mui/material/Container";
+import CssBaseline from "@mui/material/CssBaseline";
+import Grid from "@mui/material/Grid";
+import LockOutlinedIcon from "@mui/icons-material/LockOutlined";
+import TextField from "@mui/material/TextField";
+import Typography from "@mui/material/Typography";
+import { createTheme, ThemeProvider } from "@mui/material/styles";
+
+import { funcSubmit } from "../../components/userAuth/Submit";
+
 
 const theme = createTheme();
 
 export default function SignUp() {
   const navigate = useNavigate();
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    const data = new FormData(event.currentTarget);
-    interface Obj {
-      email: any;
-      password: any;
-    }
-    const user: Obj = {
-      email: data.get("email"),
-      password: data.get("password"),
-    };
-    try {
-      await auth.createUserWithEmailAndPassword(user.email, user.password);
+    const authResult = await funcSubmit(event, "signup");
+    if (authResult) {
       navigate("../mypage/meeting");
-    } catch (error) {
-      console.error(error);
     }
   };
 
@@ -105,12 +83,11 @@ export default function SignUp() {
             </Button>
             <Grid container justifyContent="flex-end">
               <Grid item>
-                <Link to="/signin">Already have an account? Sign in</Link>
+                <Link to="/signin">もうすでにアカウントを持っている場合 Sign in</Link>
               </Grid>
             </Grid>
           </Box>
         </Box>
-        {/* <Copyright sx={{ mt: 5 }} /> */}
       </Container>
     </ThemeProvider>
   );
